put to float

+++ openpifpaf_detection_attributes/datasets/generators.py
@@ -44,7 +44,7 @@ class BoxAttributeGenerator(AttributeGenerator):
             (n_targets, width_height[1], width_height[0]),
             init_value,
             dtype=np.float32,
-        )
+        ).astype(np.float)
         self.previous_distances = np.full((width_height[1], width_height[0]),
                                           np.inf, dtype=np.float32)
         self.previous_bottoms = np.full((width_height[1], width_height[0]),
@@ -239,7 +239,7 @@ class BoxGaussianAttributeGenerator(AttributeGenerator):
             y, x = np.arange(self.targets.shape[1]), np.arange(self.targets.shape[2])    
             gy = np.exp(-(y-y0)**2/(2*sigma_y**2))
             gx = np.exp(-(x-x0)**2/(2*sigma_x**2))
-            g  = np.outer(gy, gx)
+            g  = np.outer(gy, gx).astype(np.float)
             self.targets = np.amax([self.targets, np.expand_dims(g, axis=0)], axis=0) 
 
         else:
